requirements.txt
flask==3.0.2
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-babel>=4.0.0
pymysql>=1.1.0
opencv-python>=4.9.0.80
numpy>=2.0.0
mediapipe>=0.10.31
rapidocr_onnxruntime>=1.3.0